workshop/speech-to-text/requirements.txt
openai
google-genai